+++ src/Ydb.Sdk/src/Ado/YdbException.cs
@@ -18,7 +18,6 @@ public class YdbException : DbException
     internal static YdbException FromServer(StatusIds.Types.StatusCode statusCode, IReadOnlyList<IssueMessage> issues)
     {
         var code = statusCode.Code();
-
         var message = code.ToMessage(issues);
 
         return new YdbException(code, message);
@@ -28,10 +27,8 @@ public class YdbException : DbException
         : base(message, innerException)
     {
         Code = statusCode;
-        var policy = RetrySettings.DefaultInstance.GetRetryRule(statusCode).Policy;
-
-        IsTransient = policy == RetryPolicy.Unconditional;
-        IsTransientWhenIdempotent = policy != RetryPolicy.None;
+        IsTransient = statusCode.IsTransient();
+        IsTransientWhenIdempotent = statusCode.IsTransientWhenIdempotent();
         // TODO: Add SQLSTATE message with order with https://en.wikipedia.org/wiki/SQLSTATE
     }